Eliminate controller files for suggestion modals (#6859)
* Removed seperate controllers

* Fix lint

* Fix e2e

* Removed Show*

* Change names

* Fix lint
