+++ core/templates/dev/head/pages/creator_dashboard/CreatorDashboard.ts
@@ -137,7 +137,7 @@ require(
   'TopicsAndSkillsDashboardBackendApiService.ts'
 );
 require('pages/exploration_editor/feedback_tab/ThreadStatusDisplayService.ts');
-require('pages/suggestion_editor/ShowSuggestionModalForCreatorViewService.ts');
+require('pages/suggestion_editor/SuggestionModalForCreatorDashboardService.ts');
 require('domain/utilities/UrlInterpolationService.ts');
 require('services/AlertsService.ts');
 require('services/DateTimeFormatService.ts');
@@ -177,7 +177,7 @@ oppia.controller('CreatorDashboard', [
   '$http', '$log', '$q', '$rootScope', '$scope', '$uibModal', '$window',
   'AlertsService', 'CreatorDashboardBackendApiService', 'DateTimeFormatService',
   'ExplorationCreationService', 'QuestionObjectFactory',
-  'RatingComputationService', 'ShowSuggestionModalForCreatorViewService',
+  'RatingComputationService', 'SuggestionModalForCreatorDashboardService',
   'SuggestionObjectFactory', 'SuggestionThreadObjectFactory',
   'ThreadStatusDisplayService', 'TopicsAndSkillsDashboardBackendApiService',
   'UrlInterpolationService', 'UserService', 'EXPLORATIONS_SORT_BY_KEYS',
@@ -188,7 +188,7 @@ oppia.controller('CreatorDashboard', [
       $http, $log, $q, $rootScope, $scope, $uibModal, $window,
       AlertsService, CreatorDashboardBackendApiService, DateTimeFormatService,
       ExplorationCreationService, QuestionObjectFactory,
-      RatingComputationService, ShowSuggestionModalForCreatorViewService,
+      RatingComputationService, SuggestionModalForCreatorDashboardService,
       SuggestionObjectFactory, SuggestionThreadObjectFactory,
       ThreadStatusDisplayService, TopicsAndSkillsDashboardBackendApiService,
       UrlInterpolationService, UserService, EXPLORATIONS_SORT_BY_KEYS,
@@ -448,7 +448,7 @@ oppia.controller('CreatorDashboard', [
     };
 
     $scope.showSuggestionModal = function() {
-      ShowSuggestionModalForCreatorViewService.showSuggestionModal(
+      SuggestionModalForCreatorDashboardService.showSuggestionModal(
         $scope.activeThread.suggestion.suggestionType,
         {
           activeThread: $scope.activeThread,
