+++ core/templates/dev/head/pages/learner_dashboard/LearnerDashboard.ts
@@ -30,7 +30,7 @@ require('domain/feedback_message/FeedbackMessageSummaryObjectFactory.ts');
 require('domain/feedback_thread/FeedbackThreadSummaryObjectFactory.ts');
 require('domain/learner_dashboard/LearnerDashboardBackendApiService.ts');
 require('pages/exploration_editor/feedback_tab/ThreadStatusDisplayService.ts');
-require('pages/suggestion_editor/ShowSuggestionModalForLearnerViewService.ts');
+require('pages/suggestion_editor/SuggestionModalForLearnerDashboardService.ts');
 require('domain/utilities/UrlInterpolationService.ts');
 require('services/AlertsService.ts');
 require('services/DateTimeFormatService.ts');
@@ -95,7 +95,7 @@ oppia.controller('LearnerDashboard', [
   'LEARNER_DASHBOARD_SUBSECTION_I18N_IDS', 'ThreadStatusDisplayService',
   'DateTimeFormatService', 'FEEDBACK_THREADS_SORT_BY_KEYS_AND_I18N_IDS',
   'FeedbackThreadSummaryObjectFactory', 'FeedbackMessageSummaryObjectFactory',
-  'ShowSuggestionModalForLearnerViewService', 'UserService',
+  'SuggestionModalForLearnerDashboardService', 'UserService',
   function(
       $scope, $rootScope, $q, $window, $http, $uibModal,
       AlertsService, EXPLORATIONS_SORT_BY_KEYS_AND_I18N_IDS,
@@ -105,7 +105,7 @@ oppia.controller('LearnerDashboard', [
       LEARNER_DASHBOARD_SUBSECTION_I18N_IDS, ThreadStatusDisplayService,
       DateTimeFormatService, FEEDBACK_THREADS_SORT_BY_KEYS_AND_I18N_IDS,
       FeedbackThreadSummaryObjectFactory, FeedbackMessageSummaryObjectFactory,
-      ShowSuggestionModalForLearnerViewService, UserService) {
+      SuggestionModalForLearnerDashboardService, UserService) {
     $scope.EXPLORATIONS_SORT_BY_KEYS_AND_I18N_IDS = (
       EXPLORATIONS_SORT_BY_KEYS_AND_I18N_IDS);
     $scope.SUBSCRIPTION_SORT_BY_KEYS_AND_I18N_IDS = (
@@ -488,7 +488,7 @@ oppia.controller('LearnerDashboard', [
     };
 
     $scope.showSuggestionModal = function(newContent, oldContent, description) {
-      ShowSuggestionModalForLearnerViewService.showSuggestionModal(
+      SuggestionModalForLearnerDashboardService.showSuggestionModal(
         'edit_exploration_state_content',
         {
           newContent: newContent,
