--- core/templates/dev/head/pages/suggestion_editor/ShowSuggestionModalForLearnerView.ts
@@ -1,33 +1,0 @@
-// Copyright 2019 The Oppia Authors. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS-IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-/**
- * @fileoverview Controller to show suggestion modal in learner view.
- */
-
-require('pages/suggestion_editor/SuggestionModalService.ts');
-
-oppia.controller('ShowSuggestionModalForLearnerView', [
-  '$scope', '$uibModalInstance', 'SuggestionModalService', 'description',
-  'newContent', 'oldContent',
-  function($scope, $uibModalInstance, SuggestionModalService, description,
-      newContent, oldContent) {
-    $scope.newContent = newContent;
-    $scope.oldContent = oldContent;
-    $scope.description = description;
-    $scope.cancel = function() {
-      SuggestionModalService.cancelSuggestion($uibModalInstance);
-    };
-  }
-]);
