+++ core/templates/dev/head/pages/suggestion_editor/SuggestionModalForExplorationEditorService.ts
@@ -21,10 +21,10 @@ require('domain/utilities/UrlInterpolationService.ts');
 require('pages/exploration_editor/ExplorationDataService.ts');
 require('pages/exploration_editor/ExplorationStatesService.ts');
 require('pages/exploration_editor/feedback_tab/ThreadDataService.ts');
-require('pages/suggestion_editor/ShowSuggestionModalForEditorView.ts');
 require('pages/suggestion_editor/SuggestionModalService.ts');
+require('services/EditabilityService.ts');
 
-oppia.factory('ShowSuggestionModalForEditorViewService', [
+oppia.factory('SuggestionModalForExplorationEditorService', [
   '$log', '$rootScope', '$uibModal',
   'ExplorationDataService', 'ExplorationStatesService',
   'StateObjectFactory', 'SuggestionModalService',
@@ -35,7 +35,7 @@ oppia.factory('ShowSuggestionModalForEditorViewService', [
       ThreadDataService, UrlInterpolationService) {
     var _templateUrl = UrlInterpolationService.getDirectiveTemplateUrl(
       '/pages/suggestion_editor/' +
-      'editor_view_suggestion_modal_directive.html'
+      'exploration_editor_suggestion_modal_directive.html'
     );
 
     var _showEditStateContentSuggestionModal = function(
@@ -70,7 +70,74 @@ oppia.factory('ShowSuggestionModalForEditorViewService', [
             return activeThread.getReplacementHtmlFromSuggestion();
           }
         },
-        controller: 'ShowSuggestionModalForEditorView'
+        controller: [
+          '$log', '$scope', '$uibModalInstance', 'EditabilityService',
+          'SuggestionModalService', 'currentContent', 'description',
+          'newContent', 'suggestionIsHandled', 'suggestionIsValid',
+          'suggestionStatus', 'unsavedChangesExist',
+          function(
+              $log, $scope, $uibModalInstance, EditabilityService,
+              SuggestionModalService, currentContent, description,
+              newContent, suggestionIsHandled, suggestionIsValid,
+              suggestionStatus, unsavedChangesExist) {
+            $scope.isNotHandled = !suggestionIsHandled;
+            $scope.canEdit = EditabilityService.isEditable();
+            $scope.commitMessage = '';
+            $scope.reviewMessage = '';
+            $scope.canReject = $scope.canEdit && $scope.isNotHandled;
+            $scope.canAccept = $scope.canEdit && $scope.isNotHandled &&
+              suggestionIsValid && !unsavedChangesExist;
+
+            if (!$scope.canEdit) {
+              $scope.errorMessage = '';
+            } else if (!$scope.isNotHandled) {
+              $scope.errorMessage = ((suggestionStatus === 'accepted' ||
+                suggestionStatus === 'fixed') ?
+              SuggestionModalService.SUGGESTION_ACCEPTED_MSG :
+              SuggestionModalService.SUGGESTION_REJECTED_MSG);
+            } else if (!suggestionIsValid) {
+              $scope.errorMessage = SuggestionModalService
+                .SUGGESTION_INVALID_MSG;
+            } else if (unsavedChangesExist) {
+              $scope.errorMessage = SuggestionModalService.UNSAVED_CHANGES_MSG;
+            } else {
+              $scope.errorMessage = '';
+            }
+
+            $scope.currentContent = currentContent;
+            $scope.newContent = newContent;
+
+            $scope.acceptSuggestion = function() {
+              SuggestionModalService.acceptSuggestion(
+                $uibModalInstance,
+                {
+                  action: SuggestionModalService.ACTION_ACCEPT_SUGGESTION,
+                  commitMessage: $scope.commitMessage,
+                  reviewMessage: $scope.reviewMessage,
+                  // TODO(sll): If audio files exist for the content being
+                  // replaced, implement functionality in the modal for the
+                  // exploration creator to indicate whether this change
+                  // requires the corresponding audio subtitles to be updated.
+                  // For now, we default to assuming that the changes are
+                  // sufficiently small as to warrant no updates.
+                  audioUpdateRequired: false
+                });
+            };
+
+            $scope.rejectSuggestion = function() {
+              SuggestionModalService.rejectSuggestion(
+                $uibModalInstance,
+                {
+                  action: SuggestionModalService.ACTION_REJECT_SUGGESTION,
+                  reviewMessage: $scope.reviewMessage
+                });
+            };
+
+            $scope.cancelReview = function() {
+              SuggestionModalService.cancelSuggestion($uibModalInstance);
+            };
+          }
+        ]
       }).result.then(function(result) {
         ThreadDataService.resolveSuggestion(
           activeThread.threadId, result.action, result.commitMessage,
